requirements.txt
streamlit
requests
flask
openai
python-dotenv